Restructure planner code
This patch moves the apache licensed planner code into it's own
subdirectory.
