--- src/plan_agg_bookend.h
@@ -1,10 +1,0 @@
-/*
- * This file and its contents are licensed under the Apache License 2.0.
- * Please see the included NOTICE for copyright information and
- * LICENSE-APACHE for a copy of the license.
- */
-#ifndef TIMESCALEDB_PLAN_AGG_BOOKEND_H
-#define TIMESCALEDB_PLAN_AGG_BOOKEND_H
-
-extern void ts_preprocess_first_last_aggregates(PlannerInfo *root, List *tlist);
-#endif /* TIMESCALEDB_PLAN_AGG_BOOKEND_H */
