+++ src/planner/expand_hypertable.c
@@ -3,6 +3,23 @@
  * Please see the included NOTICE for copyright information and
  * LICENSE-APACHE for a copy of the license.
  */
+
+/* This planner optimization reduces planning times when a hypertable has many chunks.
+ * It does this by expanding hypertable chunks manually, eliding the `expand_inherited_tables`
+ * logic used by PG.
+ *
+ * Slow planning time were previously seen because `expand_inherited_tables` expands all chunks of
+ * a hypertable, without regard to constraints present in the query. Then, `get_relation_info` is
+ * called on all chunks before constraint exclusion. Getting the statistics on many chunks ends
+ * up being expensive because RelationGetNumberOfBlocks has to open the file for each relation.
+ * This gets even worse under high concurrency.
+ *
+ * This logic solves this by expanding only the chunks needed to fulfil the query instead of all
+ * chunks. In effect, it moves chunk exclusion up in the planning process. But, we actually don't
+ * use constraint exclusion here, but rather a variant of range exclusion implemented by
+ * HypertableRestrictInfo.
+ * */
+
 #include <postgres.h>
 #include <catalog/pg_constraint.h>
 #include <catalog/pg_inherits.h>
@@ -36,8 +53,7 @@
 #include "import/planner.h"
 #include "nodes/chunk_append/chunk_append.h"
 #include "partitioning.h"
-#include "plan_expand_hypertable.h"
-#include "plan_partialize.h"
+#include "partialize.h"
 #include "planner.h"
 #include "time_utils.h"
